+++ backend/testsuite/backend_integration_test.go
@@ -63,31 +63,6 @@ func (s *vfsTestSuite) SetupSuite() {
 		case "sftp":
 			s.testLocations[l.FileSystem().Scheme()] = l.(*sftp.Location)
 		case "gs":
-			l.FileSystem().(*gs.FileSystem).
-				WithOptions(
-					gs.Options{
-						Retry: func(wrapped func() error) error {
-							var retryErr error
-							for i := 0; i < 5; i++ {
-								if err := wrapped(); err != nil {
-									// skip retrying for exists check
-									if err.Error() == "storage: object doesn't exist" {
-										retryErr = err
-										break
-									}
-									fmt.Printf("Initial GCS request failed. Retry (%d)\n", i+1)
-									retryErr = err
-									time.Sleep(3 * time.Second)
-									continue
-								} else {
-									retryErr = nil
-									break
-								}
-							}
-							return retryErr
-						},
-					},
-				)
 			s.testLocations[l.FileSystem().Scheme()] = l.(*gs.Location)
 		case "mem":
 			s.testLocations[l.FileSystem().Scheme()] = l.(*mem.Location)
@@ -939,7 +914,7 @@ func (s *vfsTestSuite) gsList(baseLoc vfs.Location) {
 
 	files, err := f.Location().List()
 	s.NoError(err)
-	s.Len(len(files), 1, "check file count found")
+	s.Len(files, 1, "check file count found")
 	s.Equal("file.txt", files[0], "file.txt was found")
 
 	// CLEAN UP